[backend/style] branch conflict 제거 및 폴더 구조 변경
- 실행 후 application.yml / build.gradle 수정해야 함
- 유닛 테스트 통과
- common 디렉터리로 인증 및 설정 관련 디렉터리 옮김 (필요하면 추가 작업 가능)
