+++ backend/.gitignore
@@ -5,7 +5,7 @@ build/
 !**/src/main/**/build/
 !**/src/test/**/build/
 
-.application-private.properties
+**/.application-private.properties
 
 ### STS ###
 .apt_generated
