+++ backend/src/main/java/km/cd/backend/user/UserController.java
@@ -1,6 +1,6 @@
 package km.cd.backend.user;
 
-import km.cd.backend.jwt.PrincipalDetails;
+import km.cd.backend.common.jwt.PrincipalDetails;
 import km.cd.backend.user.dto.UserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
