--- backend/src/main/resources/application-private.properties
@@ -1,17 +1,0 @@
-### JPA
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto=create
-spring.output.ansi.enabled=always
-
-# H2 console
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
-
-# H2 Database
-spring.datasource.url=jdbc:h2:~/test;
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
